+++ macro/hvenh.py
@@ -20,7 +20,7 @@ class HazardousValley(Dungeon):
   selected_dungeon = "Hazardous Valley (Easy)"
   dungeon_list = [
     "Hazardous Valley (Hard)",
-    "Hazardous Valley (Medium)",
+    "Hazardous Valley (Normal)",
     "Hazardous Valley (Easy)"
   ]
 
@@ -49,7 +49,7 @@ class HazardousValley(Dungeon):
 
       # Click Cabal Window
       util.go_cabal_window()
-      util.release_keys()
+      util.do_key_release()
       util.go_skill_slot(0.2)
       util.do_buffs()
 
@@ -76,35 +76,54 @@ class HazardousValley(Dungeon):
       util.move_scroll(250, 150, 700, 150, 0.3)
 
       util.move(500, 300)
-      util.do_dash(0.5)
-      util.wait(3)
-      atk.attack_monsters(consts.UNIT_CUTTER_TOAD, True, 0.3, self.val_sidestep_disabled)
+      if util.get_attack_type() == consts.IS_MELEE: util.do_dash(3.5)
+      else: util.do_dash()
+      atk.attack_monsters(consts.UNIT_CUTTER_TOAD, False, 0.3, self.val_sidestep_disabled)
 
       util.move(375, 150)
       util.do_dash()
-      util.do_fade(1.5)
+      atk.attack_monsters(consts.UNIT_CUTTER_TOAD, False, 0.3, self.val_sidestep_disabled)
+
+      if util.get_attack_type() == consts.IS_MELEE: util.do_dash(5.5)
+      else: util.do_dash()
       util.do_fade()
-      util.wait(6)
-      atk.attack_monsters(consts.UNIT_CUTTER_TOAD, True, 0.3, self.val_sidestep_disabled)
+      atk.attack_monsters(consts.UNIT_CUTTER_TOAD, False, 0.3, self.val_sidestep_disabled)
 
       util.move(375, 175)
       util.do_dash()
       util.do_fade()
       util.do_dash()
-      util.do_fade(1)
+      if util.get_attack_type() == consts.IS_MELEE:
+        util.move(375, 175)
+        util.do_fade(1)
 
-      util.move(250, 350)
-      util.do_dash(3)
-      util.do_fade(2)
-      atk.attack_monsters(consts.UNIT_CUTTER_TOAD, True, 0.3, self.val_sidestep_disabled)
+        util.move(250, 350)
+        util.do_dash(3)
+        util.do_fade(2)
+      else:
+        util.move(375, 175)
+        util.do_fade()
+
+        util.move(250, 350)
+        util.do_dash()
+        util.do_fade()
+      util.do_final_mode(1)
+      atk.attack_monsters(consts.UNIT_CUTTER_TOAD, False, 0.3, self.val_sidestep_disabled)
 
       util.move(535, 150)
       util.do_dash()
       util.do_fade()
-
       util.move(400, 200)
       util.do_dash()
-      util.do_fade(1.5)
+
+      atk.attack_monsters(consts.UNIT_CUTTER_TOAD, False, 0.3, self.val_sidestep_disabled)
+
+      util.do_dash_rel(456, 156)
+      util.do_fade_rel(416, 253)
+      util.do_dash_rel(434, 333)
+
+      atk.attack_monsters(consts.UNIT_CUTTER_TOAD, False, 0.3, self.val_sidestep_disabled)
+      util.do_fade_rel(533, 327)
 
       # Check Macro State
       if not util.get_macro_state():
@@ -139,7 +158,7 @@ class HazardousValley(Dungeon):
       util.move(630, 520)
       util.do_fade()
       util.move(550, 250)
-      util.do_dash(0.5)
+      util.do_dash(0.1)
 
       checking = True
       while checking: