+++ macro/hvv.py
@@ -15,6 +15,29 @@ import common.attack as atk
 pynboard = Controller()
 
 class HazardousValleyVeradrix(Dungeon):
+  DG_HVV = ""
+
+  # UNIQUE VARIABLES
+  selected_dungeon = "Hazardous Valley (Veradrix - Easy)"
+  dungeon_list = [
+    "Hazardous Valley (Veradrix - Hard)",
+    "Hazardous Valley (Veradrix - Normal)",
+    "Hazardous Valley (Veradrix - Easy)"
+  ]
+
+  def initialize(self, args):
+    self.frame_root = args[consts.DATA_FRAME]
+    self.btn_start = args[consts.DATA_BUTTON]
+    self.selected_dungeon = args[consts.DATA_DUNGEON]
+
+    shortcut.add_hotkey(consts.HOTKEY_TERMINATE, util.terminate)
+    self.btn_start.config(state=consts.STATE_DISABLED)
+    self.frame_root.update()
+
+    self.run_dungeon(args[consts.DATA_RUNS])
+
+    self.btn_start.config(state=consts.STATE_NORMAL)
+    self.frame_root.update()
 
   def run_dungeon(self, runs):
     run_counter = 0
@@ -27,7 +50,7 @@ class HazardousValleyVeradrix(Dungeon):
 
       # Click Cabal Window
       util.go_cabal_window()
-      util.release_keys()
+      util.do_key_release()
       util.go_skill_slot(0.2)
       util.do_buffs()
 
@@ -40,7 +63,13 @@ class HazardousValleyVeradrix(Dungeon):
       util.move(677, 361)
       util.move(735, 361, 0.5)
       self.click_dungeon_portal(735, 361)
-      util.move_click(440, 300, 0.5)
+
+      if self.selected_dungeon == self.dungeon_list[0]:
+        util.move_click(440, 300, 0.5)
+      elif self.selected_dungeon == self.dungeon_list[1]:
+        util.move_click(440, 280, 0.5)
+      elif self.selected_dungeon == self.dungeon_list[2]:
+        util.move_click(440, 260, 0.5)
 
       # Enter Dungeon
       self.enter_dungeon()
@@ -157,9 +186,9 @@ class HazardousValleyVeradrix(Dungeon):
         run_counter += 1000
         continue
 
-      util.move(570, 260)
+      util.move_click(570, 260)
       util.do_fade()
-      util.do_dash()
+      util.do_dash(0.5)
 
       atk.plunder_box()